requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-rest-passwordreset==1.2.1
python-dotenv==0.20.0
django-import-export==2.8.0
django-admin-interface==0.19.0